requirements.txt
streamlit>=1.30.0
python-docx>=1.1.0
google-genai>=1.0.0
anthropic>=0.42.0
openai>=1.60.0
olefile>=0.47
python-hwpx>=2.9.0
